+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # Changelog
 
 ### v0.3.0
+- :bulb: Added `NumEmailSent` score criteria
+- :hammer: Fixed `ZazoActivity` score criteria (not it is just connections count)